+++ src/main/java/com/github/tartaricacid/touhoulittlemaid/util/MaidFluidRender.java
@@ -41,7 +41,7 @@ public final class MaidFluidRender {
         return FluidVariantAttributes.getName(FluidVariant.of(fluid));
     }
 
-    public static void drawFluid(GuiGraphics graphics, int x, int y, int width, int height, String fluidId, int amount, int capacity) {
+    public static void drawFluid(GuiGraphics graphics, int x, int y, int width, int height, String fluidId, long amount, long capacity) {
         ResourceLocation location = ResourceLocation.tryParse(fluidId);
         Fluid fluid = BuiltInRegistries.FLUID.get(location);
         if (location == null || amount <= 0 || fluid.isSame(Fluids.EMPTY)) {
@@ -50,7 +50,7 @@ public final class MaidFluidRender {
         FluidVariant fluidStack = FluidVariant.of(fluid);
         getStillFluidSprite(fluidStack).ifPresent(fluidStillSprite -> {
             int fluidColor = getColorTint(fluidStack);
-            int scaledAmount = (amount * height) / capacity;
+            long scaledAmount = (amount * height) / capacity;
             if (scaledAmount < 1) {
                 // 至少渲染一行像素，让人知道里面有东西
                 scaledAmount = 1;
